requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
faiss-cpu
sentence-transformers
pypdf2
pdf2image
pytesseract
